feature/VIH-6671 : refactoring reduce number of parameters
